backend/requirements-observability.txt
# Core Framework
fastapi==0.116.1
starlette==0.47.2

# Typing & Validation
pydantic==2.11.7
annotated-types==0.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1

# Async & Networking
anyio==4.10.0
sniffio==1.3.1
urllib3==2.5.0
certifi==2025.8.3

# Logging & Monitoring
structlog
psutil==6.1.0
wrapt==1.17.3
wcwidth==0.2.13

# Error Tracking
sentry-sdk==2.35.0

# OpenTelemetry Core
deprecated==1.2.14
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-sqlalchemy==0.57b0
